feat: Enable the cluster module to support fallback encryption keys
