+++ agent/src/cluster/mod.rs
@@ -4,8 +4,10 @@ mod node;
 mod versioned;
 mod store;
 mod transport;
+mod encryption;
 
 pub use client::*;
+pub use encryption::*;
 pub use message::*;
 pub use node::*;
 pub use versioned::*;
